requirements.txt
yfinance>=0.2.36
pandas>=2.0
numpy>=1.24
requests>=2.28
openai>=1.0
anthropic>=0.20
python-dotenv>=1.0
jinja2>=3.1
tiktoken>=0.5
pyyaml>=6.0
streamlit>=1.33
